test(BA-6979): keep the replica-group history cases in their own file
TestSchedulingHistoryRepository goes back to what it was; the
replica-group searches get their own module with the endpoint and
replica-group setup they need.
